% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilution_calculation.R
\name{dilution_calculation}
\alias{dilution_calculation}
\title{Dilution calculator}
\usage{
dilution_calculation(data_table, concentration_column,
  target_concentration = 2, maximum_sample_volume = 40,
  maximum_volume = 200, print_for_robot = FALSE,
  output_directory = getwd(), output_filename = NULL)
}
\arguments{
\item{data_table}{data table containing smples to be diluted}

\item{concentration_column}{string containing name of column that contains concentration information}

\item{target_concentration}{target concentration of dilution (e.g. 2 for 16S, 25 for metagenomics)}

\item{maximum_sample_volume}{maximum amount of sample to be used to generate diluted sample (should definitely be less than the volume of your sample, but if it is too small, will limit the number of your samples that can be accurately diluted)}

\item{maximum_volume}{maximum volume of diluted sample (often limited by plate volume, e.g. 200 uL)}

\item{print_for_robot}{TRUE/FALSE to indicate whether or not a file should be made to be used with the robot (default = FALSE)}

\item{output_directory}{directory to save file for use with the robot print_for_robot must be TRUE (default = working directory)}

\item{output_filename}{filename for output file. If none is given, default will be "robot_dilution_file_YYYY_MM_DD.csv". It is not necessary to provide the '.csv' extension.}
}
\value{
data table with additional columns that are ready for dilutions
}
\description{
Dilution calculator
}
