% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\docType{data}
\name{EJC_colors}
\alias{EJC_colors}
\title{Color Scheme function for plotting}
\format{An object of class \code{character} of length 2300.}
\usage{
EJC_colors
}
\description{
Color Scheme function for plotting
}
\keyword{datasets}
