% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{EJC_theme}
\alias{EJC_theme}
\title{Theme function for plotting}
\usage{
EJC_theme(base_size = 12, base_family = "sans", bgcolor = "default")
}
\arguments{
\item{base_size}{base text size}

\item{base_family}{base text family}

\item{bgcolor}{background color}
}
\description{
Theme function for plotting
}
