% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{EJC_theme_tilted}
\alias{EJC_theme_tilted}
\title{Theme function for plotting with tilted axes}
\usage{
EJC_theme_tilted(base_size = 12, base_family = "sans",
  bgcolor = "default")
}
\arguments{
\item{base_size}{base text size}

\item{base_family}{base text family}

\item{bgcolor}{background color}
}
\description{
Theme function for plotting with tilted axes
}
