% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Example_Bruker_biotyper_html_file}
\alias{Example_Bruker_biotyper_html_file}
\title{Bruker Biotyper sample output html file}
\format{HTML file}
\source{
Generated in lab.
}
\usage{
Example_Bruker_biotyper_html_file
}
\description{
HTML file output from bruker biotyper
}
\keyword{datasets}
