% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Example_empty_weights}
\alias{Example_empty_weights}
\title{Empty tube weight file}
\format{tab-delimited text file}
\source{
Generated in lab for the purposes of analyzing this type of data
}
\usage{
Example_empty_weights
}
\description{
A file that contains the empty tube weights of tubes used in experiment - scanned for tube barcode and BarcodeID for linking with mapping file
}
\keyword{datasets}
