% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Example_full_weights}
\alias{Example_full_weights}
\title{Full tube weight file}
\format{tab-delimited text file}
\source{
Generated in lab for the purposes of analyzing this type of data
}
\usage{
Example_full_weights
}
\description{
A file that contains the full tube weights of tubes used in experiment - scanned for tube barcode for linking with empty weights file
}
\keyword{datasets}
