% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Example_HS_standard_mapping}
\alias{Example_HS_standard_mapping}
\title{Mapping file for the plate containing High-Sensitivity dye standard curve}
\format{A table with 96 rows and 10 columns}
\source{
Generated in lab for the purposes of analyzing this type of data
}
\usage{
Example_HS_standard_mapping
}
\description{
A dataset that contains the information from a mapping, and is used to re-create a template mapping file. This data is used in the sample_file_maker script to generate a sample mapping file. This dataset contains an example of how to include standards for a standard curve
}
\keyword{datasets}
