% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Example_mapping_file}
\alias{Example_mapping_file}
\title{Mapping file for samples in microbial density experiment}
\format{A table with 96 rows and 21 columns}
\source{
Generated in lab for the purposes of analyzing this type of data
}
\usage{
Example_mapping_file
}
\description{
A dataset that contains the information from a mapping, and is used to re-create a template mapping file. This data is used in the sample_file_maker script to generate a sample mapping file
}
\keyword{datasets}
