% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Example_matrix_plate_scan}
\alias{Example_matrix_plate_scan}
\title{Tube Order from 96-well format barcode scanner (.csv)}
\format{A table with up to 96 rows}
\source{
Generated in lab for the purposes of analyzing this type of data
}
\usage{
Example_matrix_plate_scan
}
\description{
A dataset that contains the location of tubes used in experiment
}
\keyword{datasets}
