% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Example_other_data}
\alias{Example_other_data}
\title{Fully processed other data (.csv)}
\format{Table with many samples}
\source{
Generated in lab for the purposes of analyzing this type of data
}
\usage{
Example_other_data
}
\description{
A dataset that contains processed data from othere experiments
}
\keyword{datasets}
