% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Example_pcr_1_data}
\alias{Example_pcr_1_data}
\title{Raw Fluorescence data for the plate containing pcr product quantifications}
\format{This data is not in a standard format since it is simply the data read in directly from a .xls file.}
\source{
Generated in lab.
}
\usage{
Example_pcr_1_data
}
\description{
A dataset that contains the information from a plate reader file, and is used to re-create a raw fluorescence file as if it came from the plate reader. This data is used in the sample_file_maker script to generate sample raw files.
}
\keyword{datasets}
