% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Example_pcr_3_map}
\alias{Example_pcr_3_map}
\title{Mapping file for the plate containing pcr product quantifications}
\format{This data is not in a standard format since it is simply the data read in directly from a .csv file.}
\source{
Generated in lab.
}
\usage{
Example_pcr_3_map
}
\description{
A dataset that contains the mapping file for a pcr product quantification. This data is used in the sample_file_maker script to generate sample raw files.
}
\keyword{datasets}
