% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Example_sample_info}
\alias{Example_sample_info}
\title{Sample Info for samples in microbial density experiment}
\format{A table with 96 rows and 21 columns}
\source{
Generated in lab for the purposes of analyzing this type of data
}
\usage{
Example_sample_info
}
\description{
A dataset that contains the information used to create a mapping file
}
\keyword{datasets}
