% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{faith_lab_theme_tilted}
\alias{faith_lab_theme_tilted}
\title{Theme function for plotting paper-ready figures with tilted axes}
\usage{
faith_lab_theme_tilted(plot_size = "small", base_family = "sans",
  base_rect_size = 0)
}
\arguments{
\item{plot_size}{"small" or "large" to designate how big to make plot font and line thickness}

\item{base_family}{to change font family to e.g. ArialMT}

\item{base_rect_size}{ggplot inherited argument}
}
\description{
Theme function for plotting paper-ready figures with tilted axes
}
