% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FaithLabTools.R
\docType{package}
\name{FaithLabTools}
\alias{FaithLabTools}
\alias{FaithLabTools-package}
\title{FaithLabTools: A package containing helper functions for use around the lab.}
\description{
FaithLabTools: A package containing helper functions for use around the lab.
}
