% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_mean_sem.R
\name{geom_mean_sem}
\alias{geom_mean_sem}
\title{New layer function that plots both the summary statistics AND the quasirandom plot from ggbeeswarm}
\usage{
geom_mean_sem(mapping = NULL, data = NULL, position = "identity", ...,
  point.size = 1, point.shape = 20, point.color = NULL,
  point.stroke = 0, line.size = 1, na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, width = 0.2, varwidth = FALSE, bandwidth = 0.5,
  nbins = NULL, method = "quasirandom", groupOnX = TRUE,
  dodge.width = 0)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{position}{position}

\item{...}{...}

\item{point.size}{point.size}

\item{point.shape}{point shape}

\item{point.color}{point.color}

\item{point.stroke}{point stroke}

\item{line.size}{line.size}

\item{na.rm}{na.rm}

\item{show.legend}{show.legend}

\item{inherit.aes}{inherit.aes}

\item{width}{width}

\item{varwidth}{varwidth}

\item{bandwidth}{bandwidth}

\item{nbins}{nbins}

\item{method}{method}

\item{groupOnX}{grouponX}

\item{dodge.width}{dodge.width}
}
\value{
plot layer
}
\description{
New layer function that plots both the summary statistics AND the quasirandom plot from ggbeeswarm
}
