% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_mean_sem.R
\name{ggname}
\alias{ggname}
\title{ggname
Helper function taken from ggplot2 -- not exported, need to re-create}
\usage{
ggname(prefix, grob)
}
\arguments{
\item{prefix}{prefix}

\item{grob}{grob}
}
\value{
ggname
}
\description{
ggname
Helper function taken from ggplot2 -- not exported, need to re-create
}
