% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{index_mapping_16S}
\alias{index_mapping_16S}
\title{Table containing 16S indexes (barcodes)}
\format{Table}
\source{
Generated for helping demultiplex 16S libraries
}
\usage{
index_mapping_16S
}
\description{
A data frame containing the location and sequencing indexes for 16S
}
\keyword{datasets}
