% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{index_mapping_metagenomics}
\alias{index_mapping_metagenomics}
\title{Table containing metagenomic sequencing indexes from IDT (barcodes)}
\format{Table}
\source{
Generated for helping demultiplex metagenomics libraries
}
\usage{
index_mapping_metagenomics
}
\description{
A data frame containing the location and sequencing indexes for metagenomics
}
\keyword{datasets}
