% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_reorder.R
\name{level_reorder}
\alias{level_reorder}
\title{Reorder factor levels}
\usage{
level_reorder(vector, order)
}
\arguments{
\item{vector}{vector to reorder}

\item{order}{order of reordered factor levels}
}
\value{
reoredered factor
}
\description{
Reorder factor levels
}
