% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_example_files.R
\name{make_example_files}
\alias{make_example_files}
\title{Function to generate sample data files to work wtih locally}
\usage{
make_example_files(directory = getwd())
}
\arguments{
\item{directory}{If given, a directory into which to create files}
}
\value{
A collection of files useful for getting started with the FaithLabTools package
}
\description{
Function to generate sample data files to work wtih locally
}
