% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_and_order.R
\name{mass_and_order}
\alias{mass_and_order}
\title{Generate sample mass and final order of tubes. Will attempt to identify BarcodeID (length 29 generated sample description) and TubeBarcodes (length 10 number found on the Matrix Tubes) in files if possible}
\usage{
mass_and_order(empty_weights, full_weights, order = NULL, plate_name = NA)
}
\arguments{
\item{empty_weights}{.txt file with barcodes and weights of empty tubes}

\item{full_weights}{.txt file with barcodes and weights of full tubes}

\item{order}{file from 96 barcode scanner that gives final order of tubes (optional. if not provided, uses the order of tubes in the full tube weight file)}

\item{plate_name}{name of plate. If not provided, will pull name from the plate scanner file (RackID)}
}
\value{
table of weights and tubes with order of samples
}
\description{
Generate sample mass and final order of tubes. Will attempt to identify BarcodeID (length 29 generated sample description) and TubeBarcodes (length 10 number found on the Matrix Tubes) in files if possible
}
