% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_plate_parser.R
\name{matrix_plate_parser}
\alias{matrix_plate_parser}
\title{Matrix plate parser}
\usage{
matrix_plate_parser(matrix_barcode_plate_scan, plate_name = NA)
}
\arguments{
\item{matrix_barcode_plate_scan}{file from matrix rack scanner}

\item{plate_name}{name of plate. If not provided, will pull name from the plate scanner file (RackID)}
}
\value{
table of tubes
}
\description{
Matrix plate parser
}
