% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_dna_concentration.R
\name{measure_dna_concentration}
\alias{measure_dna_concentration}
\title{Measure DNA concentration for using qubit to quantify DNA}
\usage{
measure_dna_concentration(plate_reader_file,
  standards_plate_reader_file = plate_reader_file, standard_wells, dye_used,
  qubit_volume = 2, elution_volume = 100, plate_size = 96,
  plate_name = "DNA_Plate", print_standard_curve = FALSE)
}
\arguments{
\item{plate_reader_file}{relative location to relevant plate reader file (accepts excel files)}

\item{standards_plate_reader_file}{relative location to the plate reader file containing the standards (if it is the same as the plate you want to quantify, it is not necessary to explicitly specify}

\item{standard_wells}{a vector containing the wells that contain the standards in increasing concentration (e.g. c("A01", "A02", "A03", "A04", "A05", "A06", "A07", "A08"))}

\item{dye_used}{"BR" or "HS" to specify quantification with HS or BR dye}

\item{qubit_volume}{volume of sample used for quantification (default = 2)}

\item{elution_volume}{volume of your sample (used to compute total DNA in sample, default = 100)}

\item{plate_size}{size of plate (default = 96, alternative is 384 - not tested yet)}

\item{plate_name}{name to give plate (helps when using multiple plates)}

\item{print_standard_curve}{TRUE or FALSE to indicate whether or not to save a .pdf image containing the standard curve}
}
\value{
table with DNA concentrations for each sample, given the input parameters
}
\description{
Measure DNA concentration for using qubit to quantify DNA
}
