% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{paper_theme_tilted}
\alias{paper_theme_tilted}
\title{Theme function for plotting paper-ready figures with tilted axes}
\usage{
paper_theme_tilted()
}
\description{
Theme function for plotting paper-ready figures with tilted axes
}
