% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plate_mapping_96_384}
\alias{plate_mapping_96_384}
\title{Table to help map four 96-well plate locations to one 384 well plate}
\format{Table}
\source{
Generated for helping shift between 4x96-well and 384-well plate formats
}
\usage{
plate_mapping_96_384
}
\description{
A data frame containing the location of each 'quadrant' of 96 well plates to one single 384-well plate
}
\keyword{datasets}
