% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qubit_merger.R
\name{qubit_merger}
\alias{qubit_merger}
\title{Function to improve accuracy of DNA measurements for samples measured with both HS and BR dye}
\usage{
qubit_merger(hs_data, br_data)
}
\arguments{
\item{hs_data}{data frame with HS dye data}

\item{br_data}{data frame with BR dye data}
}
\value{
data frame with more accurate DNA concentrations and microbial density measurements
}
\description{
Function to improve accuracy of DNA measurements for samples measured with both HS and BR dye
}
