% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bruker_file.R
\name{read_bruker_file}
\alias{read_bruker_file}
\title{Read Bruker Biotyper data into data table}
\usage{
read_bruker_file(bruker_html_file)
}
\arguments{
\item{bruker_html_file}{path to Bruker biotyper html file output}
}
\value{
data table containing Bruker biotyper data
}
\description{
Read Bruker Biotyper data into data table
}
