% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sample_info.R
\name{read_sample_info}
\alias{read_sample_info}
\title{Function to read in sample info from a file (excel or .csv)}
\usage{
read_sample_info(file_path)
}
\arguments{
\item{file_path}{string containing path to file}
}
\value{
Sample information in a data frame
}
\description{
Function to read in sample info from a file (excel or .csv)
}
