% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robot_equal_combine.R
\name{robot_equal_combine}
\alias{robot_equal_combine}
\title{Robot equal combine}
\usage{
robot_equal_combine(dataset, min_volume = 1.5, max_volume = 6,
  number_of_pools = "Auto", maximum_pool_size = 500,
  starting_plate_size = 384, print_for_robot = FALSE,
  output_directory = getwd(), output_filename = NULL)
}
\arguments{
\item{dataset}{input data set (e.g. DNA quantification of PCR products)}

\item{min_volume}{minimum volume of sample to combine}

\item{max_volume}{maximum volume of sample to combine}

\item{number_of_pools}{number of different pools to create (default = "Auto" which creates as few as possible while maintaining each pool under the maximum pool size)}

\item{maximum_pool_size}{maximum volume of any one pool}

\item{starting_plate_size}{384 or 96}

\item{print_for_robot}{TRUE/FALSE to indicate whether or not a file should be made to be used with the robot (default = FALSE)}

\item{output_directory}{directory to save file for use with the robot print_for_robot must be TRUE (default = working directory)}

\item{output_filename}{filename for output file. If none is given, default will be "robot_pooling_file_YYYY_MM_DD.csv". It is not necessary to provide the '.csv' extension.}
}
\value{
table that helps pool samples using robot (+/- saved .csv file)
}
\description{
Robot equal combine
}
