% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robot_prep_16S.R
\name{robot_prep_16S}
\alias{robot_prep_16S}
\title{Function to produce a Beckmann Robot-friendly file and table for diluting samples for 16S Sequencing}
\usage{
robot_prep_16S(dataset, n_barcode_plates)
}
\arguments{
\item{dataset}{an input dataset that includes sample volumes needed for pcr, water volume needed for pcr, dna concentrations, and sample location information.}

\item{n_barcode_plates}{the number of barcode plates available (i.e. if you have 96 unique sequencing barcodes, this would be 1)}
}
\value{
a table contianing the necessary information for the robot to handle the dilution, as well as some QC data for evaluating if there will be any errors. This will also print to terminal the names of the plates necessary (it is useful to have your plates named.)
}
\description{
Function to produce a Beckmann Robot-friendly file and table for diluting samples for 16S Sequencing
}
