% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_tools.R
\name{slides_theme_tilted}
\alias{slides_theme_tilted}
\title{Theme function for plotting presentation-ready figures with tilted axes}
\usage{
slides_theme_tilted()
}
\description{
Theme function for plotting presentation-ready figures with tilted axes
}
