% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_mean_sem.R
\name{stat_mean_sem}
\alias{stat_mean_sem}
\title{Layer function for stat}
\usage{
stat_mean_sem(mapping = NULL, data = NULL, geom = "mean_sem",
  position = "dodge", ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\arguments{
\item{mapping}{mapping}

\item{data}{dadta}

\item{geom}{geom}

\item{position}{position}

\item{...}{...}

\item{na.rm}{na.rm}

\item{show.legend}{show.legend}

\item{inherit.aes}{inherit.aes}
}
\value{
plot layer
}
\description{
Layer function for stat
}
