% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well_number_to_location.R
\name{well_number_to_location}
\alias{well_number_to_location}
\title{Well number to location. Changes well number (e.g. 1, 2, 3, ...) to a well location (A01, A02, A03, ...)}
\usage{
well_number_to_location(well_number, size = 96)
}
\arguments{
\item{well_number}{numeric}

\item{size}{size of plate (96 or 384)}
}
\value{
well location (e.g. A01, A02, A03, etc)
}
\description{
Well number to location. Changes well number (e.g. 1, 2, 3, ...) to a well location (A01, A02, A03, ...)
}
