% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well_parser.R
\name{well_parser}
\alias{well_parser}
\title{Function to toggle between a well ID and a well location number (useful for use with the Beckmann robot)}
\usage{
well_parser(well, size = 96)
}
\arguments{
\item{well}{well ID (e.g  A01 or F05)}

\item{size}{size of plate (96 or 384)}
}
\value{
A well number (e.g. 1 or 65)
}
\description{
Function to toggle between a well ID and a well location number (useful for use with the Beckmann robot)
}
